+++ test-data/src/main/java/org/ehrbase/openehr/sdk/test_data/composition/CompositionTestDataCanonicalJson.java
@@ -66,7 +66,8 @@ public enum CompositionTestDataCanonicalJson {
     IPS("Internation Patient Summary", "ips_canonical.json"),
     IPS_INVALID("IPS Invalid", "ips_invalid.json"),
     OTHER_PARTICIPATIONS("to test querying other_participations in composition", "other_participations.json"),
-    SECTION_CARDINALITY("cardinality_of_section", "cardinality_of_section__full.json");
+    SECTION_CARDINALITY("cardinality_of_section", "cardinality_of_section__full.json"),
+    EHRN_ABDM_OP_CONSULT_RECORD("Cluster with DV_MULTIMEDIA base64 encoded data", "ehrb_adbm_op_consult_record.json");
 
     private final String filename;
     private final String description;